requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
google-genai
chromadb
pypdf
beautifulsoup4
lxml
unstructured
sentence-transformers
tiktoken
python-dotenv
selenium
webdriver-manager
fastapi
uvicorn
python-multipart
requests
markdown



